Langgraph/ResumeBot/requirements.txt
langgraph
python-dotenv
ipython
gradio
requests
langchain-openai
langchain-community
langchain
chromadb
pypdf